requirements.txt
aniso8601==9.0.1
blinker==1.6.3
boto3==1.28.76
botocore==1.31.76
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
urllib3==2.0.7
Werkzeug==3.0.1
